requirements.txt
fastapi[standard]
pyjwt
passlib[bcrypt]
black
isort
pylint
pytest
pytest-cov
python-dotenv
pydantic-settings
sqlmodel